requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-seed==0.2.2
djangorestframework==3.11.0
drf-writable-nested==0.6.1
Faker==4.1.3
flake8==3.8.3
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
mccabe==0.6.1
msgpack==1.0.0
msgpack-python==0.5.6
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==2.10.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
zipp==3.1.0
zope.interface==5.1.2
